Start watching for events from last resource version
It adds a new method to the PodWatcher(EnsureLogStream) which first gets the current RV
of the PodList, and then ensure that the current pods in the namespace are tracked.

This avoid races as we get the list of pods after getting the initial RV, so we are sure
to process all the events that happened meanwhile we set up the monitoring goroutines.

Also in this way we keep streaming logs of pods that are already in the namespace if the component
get restarted.

Requires: https://github.com/SUSE/eirinix/pull/38