fix: podwatcher: only list wanted ns initially
When starting the pod watcher, we will initially track all existing
pods.  When we do so, make sure to only look for pods in the desired
namespace, instead of all of them.  This avoids the error:

pods is forbidden: User "system:serviceaccount:kubecf:eirinix" cannot
list resource "pods" in API group "" at the cluster scope