feat: allow to skip gracePeriod patching
With the new option, it's possible to skip gracePeriod injection on opi runtime apps.

This is needed to avoid mutating docker applications, which have a custom entrypoint.

Until https://github.com/cloudfoundry-incubator/eirini/issues/112 is fixed, there is no better way
to distinguish from an application staged from buildpacks and one that is coming from a docker image.